Alteracao procurar por user

+++ src/main/kotlin/com/example/appdam/backend/services/ReceitaService.kt
@@ -7,37 +7,41 @@ import org.springframework.stereotype.Service
 @Service
 class ReceitaService(private val receitaRepository: ReceitaRepository) {
 
-     // Novo método para buscar todas as receitas do banco
-    fun listarTodasReceitas(): List<Receita> {
-        return receitaRepository.findAll()
+    fun listarReceitasPorUsername(username: String): List<Receita> {
+        return receitaRepository.findByUsername(username)
     }
 
-    // Criar uma receita
     fun criarReceita(receita: Receita): Receita {
-        return receitaRepository.save(receita) // Salva a receita com o username associado
+        return receitaRepository.save(receita)
     }
 
-    // Atualizar uma receita
     fun atualizarReceita(id: Long, novaReceita: Receita): Receita {
         val receitaExistente = receitaRepository.findById(id).orElseThrow {
             IllegalArgumentException("Receita não encontrada!")
         }
 
-        // Atualiza os campos
+        if (receitaExistente.username != novaReceita.username) {
+            throw IllegalArgumentException("Você não tem permissão para atualizar esta receita!")
+        }
+
         receitaExistente.titulo = novaReceita.titulo
         receitaExistente.ingredientes = novaReceita.ingredientes
         receitaExistente.preparo = novaReceita.preparo
         receitaExistente.fotourl = novaReceita.fotourl
 
-        // Salva a receita atualizada
         return receitaRepository.save(receitaExistente)
     }
 
-    // Excluir uma receita
-    fun deletarReceita(id: Long) {
+    fun eliminarReceita(id: Long, username: String) {
         val receita = receitaRepository.findById(id).orElseThrow {
             IllegalArgumentException("Receita não encontrada!")
         }
+
+        if (receita.username != username) {
+            throw IllegalArgumentException("Você não tem permissão para eliminar esta receita!")
+        }
+
         receitaRepository.delete(receita)
     }
 }
+
